requirements.txt
fastapi==0.111.0
langchain==0.2.17
langchain-community==0.2.19
langgraph==0.1.19
langchain-openai==0.1.20
langchain-ollama==0.1.1
langchain-postgres == 0.0.12
urllib3 >= 1.25.3, < 2.1.0
# chromadb==0.5.4
pysqlite3-binary==0.5.3
psycopg[binary]
python-json-logger==2.0.7
ruamel_yaml==0.18.6
# PyJWT==2.8.0
python-json-logger==2.0.7
sqlmodel